Change first, fifth and sixth section
Rewrote first part in terms of partial bialgebras and weak multiplier
bialgebras à la Bohm, think this presentation is the clearest, does not
affect the rest of the article very much. Pretty much done up to and
including the definition of the antipode. Put also the rewritten part of
the last section on examples into the main file, i.e. the formulas for
the quantum groupoids from ergodic actions and the particular example of
dynamical quantum SU(2) together with the representation theory of its
envelopping C*-algebra. Proofs and details for this section still have
to be provided.