Section 4
Inserted the construction of the Hopf-von Neumann bimodule.
Now see better what needs to be done at the end of Section 1.
Still to be done:
(1) simplify the construction of the weights, using just the
explicit form at the end of Prop. 4.6;
(2) insert citations and references;
(3) show that  the universal C*-algebra exists using Peter-Weyl