Sections 2, 4
Section 2: Added a Corollary 2.31 on a 2-parameter-group of
automorphisms of A obtained from the characters in Theorem 2.30
summarising all relevant algebraic properties.

Section 4: Inserted the lift of the scaling group and the unitary
antipode from the algebraic to the von-Neumann level, using the
characters and 2-parameter group of Theorem 2.30 and Corollary
2.31. Needs to be completed a bit. Need to insert the (short) proof
that this fits with Enocks theory.