Sections 1, 4
Section 1, part on integrals: changed definition into a partial form,
changed terminology to simplify the presentation, inserted a Lemma on
uniqueness and proved the strong invariance conditions which are used
in Section 4. This section is essentially finished.

Section 4: Shortened proof of Prop. 4.8 on invariance of the extended
weights to make use of the "partial  form" of invariance that is now
given in Section 1; improved proof of Lemma 4.14. This section is
essentially finished except for citations and a reference to Section
2.