Section 2
In Section 2, reworked definition of corepresentations, the tensor
product and the invertibility in the case of a partial Hopf
algebra.

(1) Right from the beginning, imposed a finiteness condition on
corepresentations which implies that the underlying vector space is
seperately finitely supported. More generality is not  needed.

(2) Related the family-wise definition to a "total" form where
corepresentations are multipliers of a certain algebra.

(3) Noted that "invertibility of corepresentations in the Hopf case"
means that the corepresentation multiplier has a generalized inverse.
Added a converse: the counit condition for corepresentations follows
if the corepresentation multiplier has a generalized inverse.